ethdev: add API to convert raw timestamps to nsec
Existing ethdev functions can read/write time from/to device but
they're all related to timesync and none of them can translate a raw
counter in real time unit which is useful in a pdump application.

A new API is required because the conversion is derived from dev
clock info.

Signed-off-by: Patrick Keroulas <patrick.keroulas@radio-canada.ca>